backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
httpx>=0.25.0
apscheduler>=3.10.4
pydantic>=2.5.0
pydantic[email]>=2.5.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-multipart>=0.0.6
pythonping>=1.1.0
aiosmtplib>=3.0.0
